+++ sdk/storage/storage-blob/test/utils.spec.ts
@@ -13,6 +13,7 @@ describe("Utility Helpers", () => {
   const endpointSuffix = "core.windows.net";
   const accountName = "myaccount";
   const blobEndpoint = `${protocol}://${accountName}.blob.${endpointSuffix}`;
+  const customDomainBlobEndpoint = `${protocol}://customdomain.com`; 
   const sharedAccessSignature = "sasToken";
 
   function verifySASConnectionString(sasConnectionString: string) {
@@ -76,6 +77,49 @@ describe("Utility Helpers", () => {
       `BlobEndpoint=${blobEndpoint};
         FileEndpoint=https://storagesample.file.core.windows.net;
         SharedAccessSignature=${sharedAccessSignature}`
+    );
+  });
+
+  it("extractConnectionStringParts parses sas connection string with custom domain", async () => {
+    const sasConnectionString = `BlobEndpoint=${customDomainBlobEndpoint};
+    SharedAccessSignature=${sharedAccessSignature}`
+    const connectionStringParts = extractConnectionStringParts(sasConnectionString);
+    assert.equal(
+      "SASConnString",
+      connectionStringParts.kind,
+      "extractConnectionStringParts().kind is different than expected."
+    );
+    assert.equal(
+      customDomainBlobEndpoint,
+      connectionStringParts.url,
+      "extractConnectionStringParts().url is different than expected."
+    );
+    assert.equal(
+      '',
+      connectionStringParts.accountName,
+      "extractConnectionStringParts().accountName is different than expected."
+    );
+  });
+
+  it("extractConnectionStringParts parses sas connection string with custom domain", async () => {
+    const sasConnectionString = `BlobEndpoint=${customDomainBlobEndpoint};
+    SharedAccessSignature=${sharedAccessSignature};
+    AccountName=${accountName}`
+    const connectionStringParts = extractConnectionStringParts(sasConnectionString);
+    assert.equal(
+      "SASConnString",
+      connectionStringParts.kind,
+      "extractConnectionStringParts().kind is different than expected."
+    );
+    assert.equal(
+      customDomainBlobEndpoint,
+      connectionStringParts.url,
+      "extractConnectionStringParts().url is different than expected."
+    );
+    assert.equal(
+      accountName,
+      connectionStringParts.accountName,
+      "extractConnectionStringParts().accountName is different than expected."
     );
   });
 });