+++ sdk/storage/storage-blob/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.2.0 (Unreleased)
 
 - Added RehydratePriority to BlobProperties and BlobItemProperties.
+- Added custom domain support.
 
 ## 12.2.0-preview.1 (2020.07)